split seal from crypto crate
