+++ .gitmodules
@@ -1,3 +1,3 @@
 [submodule "thorn-seal/SEAL"]
-	path = thorn-crypto/SEAL
+	path = thorn-seal/SEAL
 	url = git@github.com:marcosfpr/SEAL.git
