+++ thorn-crypto/Cargo.toml
@@ -1,39 +1,25 @@
 [package]
-name = "thorn-seal"
+name = "thorn-crypto"
 version = "0.1.0"
 edition = "2021"
 authors = ["marcosfpr <mfprezende@gmail.com>"]
-description = "Fork of Sunscreen's thorn_seal crate containing Rust bindings for Microsoft's SEAL Fully Homomorphic Encryption (FHE) library"
+description = "thorn cryptography primitives"
 repository = "https://github.com/marcosfpr/thorn"
 readme = "../README.md"
 license-file = "../LICENSE"
 keywords = ["federated-learning", "fl", "ai", "machine-learning"]
 categories = ["science", "cryptography"]
 
+
+[badges]
+maintenance = { status = "passively-maintained" }
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[package.metadata.docs.rs]
+all-features = true
+rustdoc-args = ["--cfg", "docsrs"]
+
 [dependencies]
-serde = { workspace = true } 
-thiserror = { workspace = true }
-static_assertions = "1.1.0"
-
-[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
-link-cplusplus = "1.0.9"
-
-[build-dependencies]
-cmake = "0.1.46"
-bindgen = "0.69.4"
-
-[dev-dependencies]
-anyhow = { workspace = true } 
-serde_json = "1.0.74"
-
-[features]
-hexl = []
-transparent-ciphertexts = []
-deterministic = []
-insecure-params = []
-
-[[example]]
-name = "sum_int"
-
-[[example]]
-name = "sum_float"
+anyhow = { workspace = true }
+serde = { workspace = true }
+tracing = { workspace = true }
