+++ thorn-seal/Cargo.toml
@@ -1,0 +1,39 @@
+[package]
+name = "thorn-seal"
+version = "0.1.0"
+edition = "2021"
+authors = ["marcosfpr <mfprezende@gmail.com>"]
+description = "Rust bindings for Microsoft's SEAL Fully Homomorphic Encryption (FHE) library"
+repository = "https://github.com/marcosfpr/thorn"
+readme = "../README.md"
+license-file = "../LICENSE"
+keywords = ["federated-learning", "fl", "ai", "machine-learning"]
+categories = ["science", "cryptography"]
+
+[dependencies]
+serde = { workspace = true } 
+thiserror = { workspace = true }
+static_assertions = "1.1.0"
+
+[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
+link-cplusplus = "1.0.9"
+
+[build-dependencies]
+cmake = "0.1.46"
+bindgen = "0.69.4"
+
+[dev-dependencies]
+anyhow = { workspace = true } 
+serde_json = "1.0.74"
+
+[features]
+hexl = []
+transparent-ciphertexts = []
+deterministic = []
+insecure-params = []
+
+[[example]]
+name = "sum_int"
+
+[[example]]
+name = "sum_float"
