+++ thorn-seal/src/encoder/mod.rs
