+++ thorn-seal/src/encryptor.rs
