+++ thorn-seal/src/error.rs
