+++ thorn-seal/src/memory.rs
