+++ thorn-seal/src/plaintext.rs
