+++ thorn-seal/src/poly_array.rs
